project-f-unified-rag-mlops/requirements.txt
# Unified: Project B (RAG) + D (training/registry) + E (feedback/metrics patterns)
python-dotenv>=1.0.0
pypdf>=4.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
langchain-text-splitters>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
openai>=1.0.0
rank-bm25>=0.2.2
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
prometheus-client>=0.20.0
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.4.0
joblib>=1.3.0
sqlalchemy>=2.0.0
pydantic>=2.0.0

# Dev / CI
pytest>=8.0.0
httpx>=0.27.0
ruff>=0.8.0
